refactor : InputView 와 ResultView 패키지 위치 변경
