+++ src/main/java/step3/CarRaceApplication.java
@@ -1,5 +1,7 @@
 package step3;
 
+import step3.view.InputView;
+
 public class CarRaceApplication {
     public static void main(String[] args) {
         RandomMoveStrategy randomMoveStrategy = new RandomMoveStrategy();
